[pipes-java] fix JAR building and test dependencies
fixes https://github.com/dagster-io/community-integrations/issues/221

- the jar building does not include the dependencies anymore

- in order to run the MainTest class in the tests the application plugin is
used to build a start script and libs directory

- the test dependencies are declared separately now so that a user of the
library doesn't depend on them

- mockito was needed in the library to create a mock class. to remove
mockito from the library the PipesContext was converted to an interface
with an empty implementation instead of the mockito mock
