+++ libraries/pipes/implementations/java/build.gradle
@@ -1,5 +1,6 @@
 plugins {
-    id 'java'
+    id 'java-library'
+    id 'application'
     id 'pmd'
     id 'checkstyle'
     id "com.vanniktech.maven.publish" version "0.30.0"
@@ -55,9 +56,9 @@ dependencies {
     implementation group: 'com.fasterxml.jackson.core', name: 'jackson-databind', version: '2.18.0'
     implementation 'info.picocli:picocli:4.7.6'
     annotationProcessor 'info.picocli:picocli-codegen:4.7.6'
-    implementation 'org.junit.jupiter:junit-jupiter-api:5.10.0'
-    implementation 'org.junit.jupiter:junit-jupiter-engine:5.10.0'
-    implementation group: 'org.mockito', name: 'mockito-core', version: '4.11.0'
+    testImplementation 'org.junit.jupiter:junit-jupiter-api:5.10.0'
+    testImplementation 'org.junit.jupiter:junit-jupiter-engine:5.10.0'
+    testImplementation group: 'org.mockito', name: 'mockito-core', version: '4.11.0'
     implementation group: 'software.amazon.awssdk', name: 's3', version: '2.29.17'
 }
 
@@ -65,13 +66,34 @@ test {
     useJUnitPlatform()
 }
 
-jar {
-    from {
-        configurations.runtimeClasspath.collect { it.isDirectory() ? it : zipTree(it) } +
-            sourceSets.main.output + sourceSets.test.output
+tasks.register('testJar', Jar) {
+    archiveClassifier.set("tests")
+    from sourceSets.test.output
+}
+tasks.named('assemble') {
+    dependsOn(tasks.named('testJar'))
+}
+application {
+    mainClass.set('io.dagster.pipes.MainTest')
+}
+tasks.named('startScripts') {
+    applicationName = "mainTest"
+    classpath = files(tasks.jar.archiveFile) + files(tasks.testJar.archiveFile) + sourceSets.test.runtimeClasspath
+}
+distributions {
+    main {
+        contents {
+            // copy our test classes jar file into the install dir
+            from(tasks.named('testJar').flatMap { it.archiveFile }) {
+                into 'lib'
+            }
+            // copy the test dependencies
+            from(configurations.testRuntimeClasspath) {
+                into 'lib'
+                duplicatesStrategy = DuplicatesStrategy.EXCLUDE
+            }
+        }
     }
-    duplicatesStrategy = DuplicatesStrategy.EXCLUDE
-    dependsOn compileJava, compileTestJava, processTestResources
 }
 
 pmd {
@@ -88,7 +110,7 @@ checkstyle {
 checkstyleMain
     .exclude('**/types/**')
 
-tasks.withType(Pmd) {
+tasks.withType(Pmd).configureEach {
     // Exclude generated types files from analysis
     exclude '**/types/**'
     ruleSetFiles = files('config/pmd-ruleset.xml')
