+++ libraries/pipes/implementations/java/tests/conftest.py
@@ -4,4 +4,4 @@ import subprocess
 
 @pytest.fixture(scope="session", autouse=True)
 def built_jar():
-    subprocess.run(["./gradlew", "build"], check=True)
+    subprocess.run(["./gradlew", "build", "installDist"], check=True)
