+++ libraries/pipes/implementations/java/tests/test_pipes.py
@@ -3,8 +3,5 @@ from dagster_pipes_tests import PipesTestSuite
 
 class TestJavaPipes(PipesTestSuite):
     BASE_ARGS = [
-        "java",
-        "-cp",
-        "build/libs/dagster-pipes-java.jar",
-        "io.dagster.pipes.MainTest",
+        "build/install/dagster-pipes-java/bin/mainTest"
     ]
